Fixed false-positive status check NPIV environment
This correct the NPIV troubleshooting environment.
In the PowerVM 2.2.1.4 we can't use 'fcstat -e fcsX' to get the
status of connectivity and we was using the command
'lsdev -dev fcX -attr attach' to get satus. Normally 'al' returns
when the connection isn't ok and 'switch' when connection is ok.

But in new releases of PowerVM the status of attach as 'switch' is
not a good method to check.
Now I commiting a new version of lsnpivs.py and has double check:
uses the lsdev... and fcstat.